Fixed int16.. ASR working

+++ minimal/Main.hs
@@ -13,7 +13,7 @@ foreign import ccall unsafe "c_init_kaldi"
                c_init_kaldi :: IO (Ptr DoAsrArgs)
 
 foreign import ccall unsafe "c_main_wrap"
-               c_main_wrap :: Int -> Ptr Int32 -> IO ()
+               c_main_wrap :: Int -> Ptr Int16 -> IO ()
 
 foreign import ccall unsafe "c_doAsr"
                c_doAsr :: Ptr DoAsrArgs -> Int -> Ptr Int32 -> IO ()
@@ -29,13 +29,13 @@ main = do
 -- inMain :: FilePath -> IO ()
 inMain path = do
   maybeAudio <- importFile path
-  case maybeAudio :: Either String (Audio Int32) of
+  case maybeAudio :: Either String (Audio Int16) of
     Left s -> putStrLn $ "wav decoding error: " ++ s
     Right (Audio rate channels samples) -> do
       putStrLn $ "rate = " ++ show rate
       putStrLn $ "channels: " ++ show channels
       print (take 100 (elems samples))
-      withArrayLen (elems samples :: [Int32]) c_main_wrap
+      withArrayLen (elems samples :: [Int16]) c_main_wrap
       -- unsafeUseAsCStringLen
 
 -- withArrayLen :: Storable a => [a] -> (Int -> Ptr a -> IO b) -> IO b
